+++ pyannote/audio/core/task.py
@@ -72,9 +72,11 @@ class Specifications:
     problem: Problem
     resolution: Resolution
 
-    # chunk duration in seconds.
-    # use None for variable-length chunks
-    duration: Optional[float] = None
+    # (maximum) chunk duration in seconds
+    duration: float
+
+    # (for variable-duration tasks only) minimum chunk duration in seconds
+    min_duration: Optional[float] = None
 
     # use that many seconds on the left- and rightmost parts of each chunk
     # to warm up the model. This is mostly useful for segmentation tasks.
@@ -95,7 +97,7 @@ class Specifications:
     permutation_invariant: bool = False
 
     @cached_property
-    def powerset(self):
+    def powerset(self) -> bool:
         if self.powerset_max_classes is None:
             return False
 
@@ -117,6 +119,12 @@ class Specifications:
                 for i in range(0, self.powerset_max_classes + 1)
             )
         )
+
+    def __len__(self):
+        return 1
+
+    def __iter__(self):
+        yield self
 
 
 class TrainDataset(IterableDataset):
@@ -191,7 +199,7 @@ class Task(pl.LightningDataModule):
 
     Attributes
     ----------
-    specifications : Specifications or dict of Specifications
+    specifications : Specifications or tuple of Specifications
         Task specifications (available after `Task.setup` has been called.)
     """
 
@@ -260,7 +268,28 @@ class Task(pl.LightningDataModule):
         """
         pass
 
-    def setup(self, stage: Optional[str] = None):
+    @property
+    def specifications(self) -> Union[Specifications, Tuple[Specifications]]:
+        # setup metadata on-demand the first time specifications are requested and missing
+        if not hasattr(self, "_specifications"):
+            self.setup_metadata()
+        return self._specifications
+
+    @specifications.setter
+    def specifications(
+        self, specifications: Union[Specifications, Tuple[Specifications]]
+    ):
+        self._specifications = specifications
+
+    @property
+    def has_setup_metadata(self):
+        return getattr(self, "_has_setup_metadata", False)
+
+    @has_setup_metadata.setter
+    def has_setup_metadata(self, value: bool):
+        self._has_setup_metadata = value
+
+    def setup_metadata(self):
         """Called at the beginning of training at the very beginning of Model.setup(stage="fit")
 
         Notes
@@ -270,7 +299,10 @@ class Task(pl.LightningDataModule):
         If `specifications` attribute has not been set in `__init__`,
         `setup` is your last chance to set it.
         """
-        pass
+
+        if not self.has_setup_metadata:
+            self.setup()
+            self.has_setup_metadata = True
 
     def setup_loss_func(self):
         pass
@@ -361,6 +393,11 @@ class Task(pl.LightningDataModule):
         loss : {str: torch.tensor}
             {"loss": loss}
         """
+
+        if isinstance(self.specifications, tuple):
+            raise NotImplementedError(
+                "Default training/validation step is not implemented for multi-task."
+            )
 
         # forward pass
         y_pred = self.model(batch["X"])